--- src/main/java/TaskListIndexOutOfBoundsException.java
@@ -1,5 +1,0 @@
-public class TaskListIndexOutOfBoundsException extends EggyException {
-    public TaskListIndexOutOfBoundsException(int taskNumber, int tasksSize) {
-        super(" Task number " + taskNumber + " is out of bounds for list of " + tasksSize + " task" + (tasksSize > 1 ? "s" : "") + ".");
-    }
-}
