+++ src/main/java/eggy/command/Command.java
@@ -1,3 +1,10 @@
+package eggy.command;
+
+import eggy.exception.EggyException;
+import eggy.storage.Storage;
+import eggy.task.TaskList;
+import eggy.ui.Ui;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
