+++ src/main/java/eggy/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package eggy.command;
+
+import eggy.storage.Storage;
+import eggy.task.TaskList;
+import eggy.ui.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
