+++ src/main/java/eggy/exception/EggyException.java
@@ -1,3 +1,5 @@
+package eggy.exception;
+
 public class EggyException extends Exception {
     public EggyException(String message) {
         super("OOPS!!!" + message);
