+++ src/main/java/eggy/exception/IncompleteCommandException.java
@@ -1,0 +1,7 @@
+package eggy.exception;
+
+public class IncompleteCommandException extends EggyException {
+    public IncompleteCommandException(String commandType) {
+        super(" The eggy.task number to be " + commandType + (commandType.equals("delete") ? "d" : "ed") + " is not specified.");
+    }
+}
