+++ src/main/java/eggy/exception/IncompleteTaskException.java
@@ -1,3 +1,5 @@
+package eggy.exception;
+
 public class IncompleteTaskException extends EggyException {
     public IncompleteTaskException(String taskType) {
         super(" The description of " + (taskType.matches("^[aeiouAEIOU].*") ? "an " : "a ") + taskType +" cannot be empty.");
