+++ src/main/java/eggy/exception/InvalidCommandException.java
@@ -1,3 +1,5 @@
+package eggy.exception;
+
 public class InvalidCommandException extends EggyException {
     public InvalidCommandException() {
         super(" I'm sorry, but I don't know what that means :-(");
