+++ src/main/java/eggy/exception/LoadTasksException.java
@@ -1,3 +1,5 @@
+package eggy.exception;
+
 public class LoadTasksException extends EggyException {
     public LoadTasksException(String filePath) {
         super("Error loading tasks from " + filePath + ".");
