+++ src/main/java/eggy/exception/SaveTasksException.java
@@ -1,3 +1,5 @@
+package eggy.exception;
+
 public class SaveTasksException extends EggyException {
     public SaveTasksException(String filePath) {
         super("Error saving tasks to " + filePath + ".");
