+++ src/main/java/eggy/exception/TaskListIndexOutOfBoundsException.java
@@ -1,0 +1,7 @@
+package eggy.exception;
+
+public class TaskListIndexOutOfBoundsException extends EggyException {
+    public TaskListIndexOutOfBoundsException(int taskNumber, int tasksSize) {
+        super(" eggy.task.Task number " + taskNumber + " is out of bounds for list of " + tasksSize + " task" + (tasksSize > 1 ? "s" : "") + ".");
+    }
+}
