+++ src/main/java/eggy/task/Deadline.java
@@ -1,3 +1,5 @@
+package eggy.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
