+++ src/main/java/eggy/task/Todo.java
@@ -1,3 +1,5 @@
+package eggy.task;
+
 public class Todo extends Task {
     public Todo (String name) {
         super(name);
