+++ text-ui-test/EXPECTED.TXT
@@ -1,9 +1,9 @@
     ____________________________________________________________
-     Hello! I'm Eggy?.
+     Hello! I'm eggy.Eggy?.
      What can I do for you?
     ____________________________________________________________
     ____________________________________________________________
-     Got it. I've added this task:
+     Got it. I've added this eggy.task:
        [T][ ] borrow book
      Now you have 1 tasks in the list.
     ____________________________________________________________
@@ -12,7 +12,7 @@
      1.[T][ ] borrow book
     ____________________________________________________________
     ____________________________________________________________
-     Got it. I've added this task:
+     Got it. I've added this eggy.task:
        [D][ ] return book (by: 02 Dec 2019 06:00 PM)
      Now you have 2 tasks in the list.
     ____________________________________________________________
@@ -23,7 +23,7 @@
      OOPS!!! I'm sorry, but I don't know what that means :-(
     ____________________________________________________________
     ____________________________________________________________
-     Nice! I've marked this task as done:
+     Nice! I've marked this eggy.task as done:
        [T][X] borrow book
     ____________________________________________________________
     ____________________________________________________________
